requirements.txt
aiogram==2.25
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
fastapi==0.109.0
frozenlist==1.4.1
isort==5.13.2
openai==1.10.0
pydantic==2.6.0
pydantic_core==2.16.1
pymongo==4.6.1
python-dotenv==1.0.1
requests==2.31.0
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1

